+++ custom_components/edgeos/data_processors/base_processor.py
@@ -59,6 +59,7 @@ class BaseProcessor:
             name=device_name,
             model=self.processor_type,
             manufacturer=DEFAULT_NAME,
+            # via_device=(DEFAULT_NAME, self._hostname)
         )
 
         return device_info