+++ custom_components/edgeos/sensor.py
@@ -8,6 +8,7 @@ from homeassistant.core import HomeAssistant
 from .common.base_entity import IntegrationBaseEntity, async_setup_base_entry
 from .common.consts import ATTR_ATTRIBUTES
 from .common.entity_descriptions import IntegrationSensorEntityDescription
+from .common.enums import DeviceTypes
 from .managers.coordinator import Coordinator
 
 _LOGGER = logging.getLogger(__name__)
@@ -33,9 +34,10 @@ class IntegrationSensorEntity(IntegrationBaseEntity, SensorEntity):
         hass: HomeAssistant,
         entity_description: IntegrationSensorEntityDescription,
         coordinator: Coordinator,
+        device_type: DeviceTypes,
         item_id: str | None,
     ):
-        super().__init__(hass, entity_description, coordinator, item_id)
+        super().__init__(hass, entity_description, coordinator, device_type, item_id)
 
         self._attr_device_class = entity_description.device_class